+++ src/components/common/buttons/common/LinkButtons.tsx
@@ -1,11 +1,11 @@
 'use client'
 import { Button } from '@/components/ui/button'
-import { ILinkButtons } from '@/types/buttons-type/linkButtons-type'
 import { redirect } from 'next/navigation'
 import React, { FC } from 'react'
 import {ChevronLeft} from 'lucide-react'
+import { ILinkButton } from '@/types/buttons-type/buttons-type'
 
-const LinkButtons: FC<ILinkButtons> = ({ title, classname, link }) => {
+const LinkButtons: FC<ILinkButton> = ({ title, classname, link }) => {
   return (
     <div>
       <Button variant={'scale'} onClick={() => redirect(link)} className={`cursor-pointer rounded-xl h-[36px] flex ${classname}`}>
