+++ src/components/common/buttons/common/ViewAllButton.tsx
@@ -1,10 +1,10 @@
 'use client'
 import { Button } from '@/components/ui/button'
-import { IViewAllButton } from '@/types/buttons-type/viewAllButton-type'
+import { IButtonWithoutTitle } from '@/types/buttons-type/buttons-type'
 import { ChevronLeft } from 'lucide-react'
 import React, { FC } from 'react'
 
-const ViewAllButton: FC<IViewAllButton> = ({ classname, onclick }) => {
+const ViewAllButton: FC<IButtonWithoutTitle> = ({ classname, onclick }) => {
     return (
         <div>
             <Button variant={'scale'} onClick={onclick} className={`cursor-pointer px-4 py-2 w-[137px] h-[36px] border border-white rounded-2xl text-sm flex text-white bg-transparent ${classname}`}>
