--- src/types/buttons-type/authSubmitButton-type.ts
@@ -1,5 +1,0 @@
-export interface IAuthSubmitButton {
-    title: string;
-    classname?: string;
-    onclick?: () => void;
-}
