--- src/types/buttons-type/changephonebutton-type.ts
@@ -1,4 +1,0 @@
-export interface IChangePhoneButton {
-    classname?: string;
-    onclick?: () => void;
-}
