--- src/types/buttons-type/linkButtons-type.ts
@@ -1,5 +1,0 @@
-export interface ILinkButtons {
-    title: string;
-    classname?: string;
-    link: string;
-}
