--- src/types/buttons-type/loginButton-type.ts
@@ -1,6 +1,0 @@
-export interface ILoginButton {
-    title: string,
-    classname?: string,
-    onclick?: () => void,
-    icon: 'google' | 'apple',
-}
