--- src/types/buttons-type/sendMessageButton-type.ts
@@ -1,4 +1,0 @@
-export interface ISendMessageButton {
-    classname?: string;
-    onclick?: () => void;
-}
