--- src/types/buttons-type/submitFilter-type.ts
@@ -1,3 +1,0 @@
-export interface ISubmitFilter {
-    onclick?: () => void;
-}
