--- src/types/buttons-type/timerButton-type.ts
@@ -1,3 +1,0 @@
-export interface ITimerButton {
-    onclick?: () => void;
-}
