--- src/types/buttons-type/viewAllButton-type.ts
@@ -1,4 +1,0 @@
-export interface IViewAllButton {
-    classname?: string,
-    onclick?: () => void;
-}
